simplify(overlay-split): apply remaining codex changes
